fix calculation of gas costs

+++ neo/SmartContract/ApplicationEngine.py
@@ -355,8 +355,7 @@ class ApplicationEngine(ExecutionEngine):
         elif api == "Neo.Storage.Put":
             l1 = len(self.EvaluationStack.Peek(1).GetByteArray())
             l2 = len(self.EvaluationStack.Peek(2).GetByteArray())
-
-            return ((l1 + l2 - 1) / 1024 + 1) * 1000
+            return (int((l1 + l2 - 1) / 1024) + 1) * 1000
 
         elif api == "Neo.Storage.Delete":
             return 100
